+++ package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@quasar/extras": "^1.16.12",
+    "jose": "^5.6.3",
     "pinia": "^2.1.7",
     "quasar": "^2.16.5",
     "vue": "^3.4.29"
