+++ src/entities/license/index.ts
@@ -1,1 +1,1 @@
-export * from './model'
+export * from "./api"
