--- src/entities/license/model/index.ts
@@ -1,10 +1,0 @@
-import { getAllLicenses } from './getAllLicenses'
-import { generateLicense } from './generateLicense'
-import { getLicense } from './getLicense'
-import { getLicenseFile } from './getLicenseFile'
-import { getDigestFile } from './getDigestFIle'
-
-export type { LicenseInfo } from './generateLicense'
-export type { License } from './getLicense'
-
-export { generateLicense, getLicense, getAllLicenses, getLicenseFile, getDigestFile }
