+++ src/entities/user/api/index.ts
@@ -1,0 +1,3 @@
+export { getToken } from "./getToken";
+export type { Token } from "./getToken";
+export { getUserInfo } from "./getUserInfo";
