+++ src/entities/user/index.ts
@@ -1,1 +1,2 @@
-export * from './model'
+export { useUserStore, type UserInfo } from "./model"
+export { type Token, getToken, getUserInfo } from "./api"
